get_location separated file + siret used as default for get_data
